08-py-scheduled-jobs/requirements.txt
fastapi
jetpack-io>=0.5.1.dev202111301638297739
uvicorn
requests